Create README.md
You can find a spreadsheet of all the posts, fact-check ratings, and Facebook engagement figures [__here__](data/facebook-fact-check.csv)

The data presents the posts, fact-check ratings and Facebook engagement figures. 
These  figures were obtained from the Facebook API in October, 2016.

The aim of this project was to verify the reliability of the media depending on the media type(left-wing,right-wing, mainstream) and the source they come from.